CSP-2556 Refactor confirm add provider locations
