--- src/SFA.DAS.Roatp.CourseManagement.Web.UnitTests/Controllers/ConfirmAddProviderLocationControllerTests/CancelAddProviderLocationGetTests.cs
@@ -1,81 +1,0 @@
-﻿using System.Text.Json;
-using AutoFixture.NUnit3;
-using FluentAssertions;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.ViewFeatures;
-using Moq;
-using NUnit.Framework;
-using SFA.DAS.Roatp.CourseManagement.Domain.ApiModels;
-using SFA.DAS.Roatp.CourseManagement.Web.Controllers;
-using SFA.DAS.Roatp.CourseManagement.Web.Infrastructure;
-using SFA.DAS.Roatp.CourseManagement.Web.UnitTests.TestHelpers;
-using SFA.DAS.Testing.AutoFixture;
-
-namespace SFA.DAS.Roatp.CourseManagement.Web.UnitTests.Controllers.ConfirmAddProviderLocationControllerTests;
-public class CancelAddProviderLocationGetTests
-{
-    [Test]
-    [MoqInlineAutoData(ApprenticeshipType.Apprenticeship)]
-    [MoqInlineAutoData(ApprenticeshipType.ApprenticeshipUnit)]
-    public void When_ApprenticeshipTypeIsApprenticeshipOrApprenticeshipUnit_Then_RemovesTempData(
-        ApprenticeshipType apprenticeshipType,
-        Mock<ITempDataDictionary> tempDataMock,
-        [Greedy] ConfirmAddProviderLocationController sut,
-        AddressItem addressItem)
-    {
-        // Arrange
-        var ukprn = 12345678;
-        sut.AddDefaultContextWithUser();
-        sut.TempData = tempDataMock.Object;
-        object serialisedAddressItem = JsonSerializer.Serialize(addressItem);
-        tempDataMock.Setup(t => t.TryGetValue(TempDataKeys.SelectedTrainingVenueAddressTempDataKey, out serialisedAddressItem));
-
-        // Act
-        sut.CancelAddProviderLocation(ukprn, apprenticeshipType);
-
-        // Assert
-        tempDataMock.Verify(t => t.Remove(TempDataKeys.SelectedTrainingVenueAddressTempDataKey));
-    }
-
-    [Test, MoqAutoData]
-    public void When_ApprenticeshipTypeIsApprenticeshipUnit_Then_RedirectsToReviewYourDetails(
-        Mock<ITempDataDictionary> tempDataMock,
-        [Greedy] ConfirmAddProviderLocationController sut,
-        AddressItem addressItem)
-    {
-        // Arrange
-        var apprenticeshipType = ApprenticeshipType.ApprenticeshipUnit;
-        var ukprn = 12345678;
-        sut.AddDefaultContextWithUser();
-        sut.TempData = tempDataMock.Object;
-        object serialisedAddressItem = JsonSerializer.Serialize(addressItem);
-        tempDataMock.Setup(t => t.TryGetValue(TempDataKeys.SelectedTrainingVenueAddressTempDataKey, out serialisedAddressItem));
-
-        // Act
-        var result = sut.CancelAddProviderLocation(ukprn, apprenticeshipType) as RedirectToRouteResult;
-
-        // Assert
-        result!.RouteName.Should().Be(RouteNames.SelectShortCourseLocationOption);
-    }
-
-    [Test, MoqAutoData]
-    public void When_ApprenticeshipTypeIsApprenticeship_Then_RedirectsToGetAddStandardSelectLocationOption(
-        Mock<ITempDataDictionary> tempDataMock,
-        [Greedy] ConfirmAddProviderLocationController sut,
-        AddressItem addressItem)
-    {
-        // Arrange
-        var apprenticeshipType = ApprenticeshipType.Apprenticeship;
-        var ukprn = 12345678;
-        sut.AddDefaultContextWithUser();
-        sut.TempData = tempDataMock.Object;
-        object serialisedAddressItem = JsonSerializer.Serialize(addressItem);
-        tempDataMock.Setup(t => t.TryGetValue(TempDataKeys.SelectedTrainingVenueAddressTempDataKey, out serialisedAddressItem));
-
-        // Act
-        var result = sut.CancelAddProviderLocation(ukprn, apprenticeshipType) as RedirectToRouteResult;
-
-        // Assert
-        result!.RouteName.Should().Be(RouteNames.GetAddStandardSelectLocationOption);
-    }
-}
